cleanup: build without warnings on gcc and clang
Quiet the -Wall -Wextra set and clang's additional checks:

  - annotate the intended fall-throughs in the float-promotion switches
  - void the unused device get-pixel coordinates
  - give the zero-argument operator call a prototype, and cast LONG_MAX
    and LONG_MIN to double where a real is range-checked against them

Three were latent bugs the warnings exposed: a memory-dump pad loop
tested a fixed counter in its condition (an infinite loop if reached),
the grow log tested a filename array's address rather than its first
byte, and evalpop negated the popped object's type before comparing it
against invalidtype (0), inverting the test so that executing a null
object raised stackunderflow instead of being the no-op the PLRM
specifies (section 3.5.5: null objects are executed with no action).

+++ src/lib/xpost_op_type.c
@@ -189,7 +189,7 @@ int Scvi(Xpost_Context *ctx,
         free(t);
         return limitcheck;
     }
-    if (dbl >= LONG_MAX || dbl <= LONG_MIN){
+    if (dbl >= (double)LONG_MAX || dbl <= (double)LONG_MIN){
         free(t);
         return limitcheck;
     }
